+++ CHANGES.md
@@ -12,6 +12,9 @@ E-Maj - Change log
     removed.
   * Remove the old deprecated versions of E-Maj rollback functions that only
     returned an integer.
+  * A fourth parameter in the emaj_import_groups_configuration() functions
+    allows to specify a non default mark name when in LOGGING state groups
+    are imported.
 
 ###Bug fixes:###